requirements.txt
black==23.1.0
jinja2==3.1.2
PyYAML==6.0
requests==2.28.2
tomli==2.0.1
typer==0.9.0